+++ MessageSearch/src/com/github/searchbadger/util/Message.java
@@ -2,12 +2,12 @@ package com.github.searchbadger.util;
 
 public class Message {
 	private final Contact contact;
-	private final int Id;
-	private final int threadId;
+	private final long Id;
+	private final long threadId;
 	private final String text;
 	private final boolean isStarred;
 	
-	public Message(Contact contact, int Id, int threadId, String text, boolean isStarred){
+	public Message(Contact contact, long Id, long threadId, String text, boolean isStarred){
 		this.contact = contact;
 		this.Id = Id;
 		this.threadId = threadId;
@@ -19,11 +19,11 @@ public class Message {
 		return contact;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return Id;
 	}
 	
-	public int getThreadId() {
+	public long getThreadId() {
 		return threadId;
 	}
